* 'master' of github.com:tensorflow/tensorflow: (296 commits)
  Go: Update generated wrapper functions for TensorFlow ops.
  Move the test back to its original test file.
  Remove "`extra_py_tests_deps` dep from `scikit_learn_test` target.
  Update documentation to match current implementation.
  [tf.data] In TensorSliceDataset's iterator, cache the conversion from PartialTensorShape to TensorShape.
  Use auto when appropriate.
  Minor refactor: move unused_min / unused_max variables to the smallest scope possible.
  Use const ref for calibration kernel function.
  Limit the Hexagon delegate creation to ARM-based Android.
  Add ReadFloatFromEnvVar() util function which can read float env.
  Update header guard for calibrator.
  Minor refactor: Index batches consistently with 'b'.
  Explicitly export files needed by other packages
  Add pip package name to error message.
  Add tests for dynamic shape input w/ rank-1 tensors.
  [tfdbg2] Replace Python built-in os/glob modules with TensorFlow's own file_io
  Upsample3D for OpenCL backend.
  Add a 'none' option to let the multi-perf-option run tool fallback to the single-perf-option run behavior.
  Explicitly export files needed by other packages
  Go: Update generated wrapper functions for TensorFlow ops.
  ...

+++ tensorflow/python/ops/array_ops.py
@@ -4982,7 +4982,8 @@ def dequantize(  # pylint: disable=missing-docstring
     mode="MIN_COMBINED",
     name=None,
     axis=None,
-    narrow_range=False):
+    narrow_range=False,
+    dtype=dtypes.float32):
   if axis is None:
     axis = -1
   elif axis < 0:
@@ -4992,10 +4993,17 @@ def dequantize(  # pylint: disable=missing-docstring
 
   if axis >= 0 or narrow_range:
     return gen_array_ops.dequantize(
-        input, min_range, max_range, mode=mode, name=name,
-        narrow_range=narrow_range, axis=axis)
+        input,
+        min_range,
+        max_range,
+        mode=mode,
+        name=name,
+        narrow_range=narrow_range,
+        axis=axis,
+        dtype=dtype)
   return gen_array_ops.dequantize(
-      input, min_range, max_range, mode=mode, name=name)
+      input, min_range, max_range, mode=mode, name=name, dtype=dtype)
+
 
 dequantize.__doc__ = gen_array_ops.dequantize.__doc__
